+++ 09-crud-webUI/src/main/java/ru/isakaev/service/GenreService.java
@@ -8,7 +8,7 @@ public interface GenreService {
 
     List<Genre> getAll();
 
-    Genre getGenre(int id);
+    Genre getGenre(Long id);
 
     Genre saveGenre(String name);
